refactor(instui-cli,ui-scripts): remove ui-upgrade-scripts
simplify the upgrade code by removing the ui-upgrade-scripts package and placing their scripts in
the packages they are called from
