+++ packages/instui-cli/lib/utils/check-dependencies.js
