+++ packages/instui-cli/lib/utils/verify-package-json.js
