+++ packages/instui-cli/package.json
@@ -22,6 +22,7 @@
   "dependencies": {
     "@instructure/command-utils": "8.14.0",
     "@instructure/instui-config": "8.14.0",
+    "@instructure/pkg-utils": "8.14.0",
     "@instructure/template-app": "8.14.0",
     "@instructure/template-component": "8.14.0",
     "@instructure/template-package": "8.14.0",
@@ -30,8 +31,8 @@
     "@instructure/ui-eslint-config": "8.14.0",
     "@instructure/ui-stylelint-config": "8.14.0",
     "@instructure/ui-template-scripts": "8.14.0",
-    "@instructure/ui-upgrade-scripts": "8.14.0",
     "@instructure/ui-webpack-config": "8.14.0",
+    "depcheck": "^1",
     "semver": "^7",
     "yargs": "^16",
     "yargs-interactive": "^3.0.0"
