+++ packages/instui-config/package-lists/v9/package-list.json
@@ -1,114 +1,113 @@
 [
- "@instructure/babel-plugin-transform-imports",
- "@instructure/browserslist-config-instui",
- "@instructure/canvas-high-contrast-theme",
- "@instructure/canvas-theme",
- "@instructure/command-utils",
- "@instructure/config-loader",
- "@instructure/console",
- "@instructure/cz-lerna-changelog",
- "@instructure/debounce",
- "@instructure/emotion",
- "eslint-plugin-instructure-ui",
- "@instructure/instructure-theme",
- "@instructure/instui-cli",
- "@instructure/instui-config",
- "@instructure/pkg-utils",
- "@instructure/shared-types",
- "@instructure/template-app",
- "@instructure/template-component",
- "@instructure/template-package",
- "@instructure/ui-a11y-content",
- "@instructure/ui-a11y-utils",
- "@instructure/ui-alerts",
- "@instructure/ui-avatar",
- "@instructure/ui-axe-check",
- "@instructure/ui-babel-preset",
- "@instructure/ui-badge",
- "@instructure/ui-billboard",
- "@instructure/ui-breadcrumb",
- "@instructure/ui-buttons",
- "@instructure/ui-byline",
- "@instructure/ui-calendar",
- "@instructure/ui-checkbox",
- "@instructure/ui-code-editor",
- "@instructure/ui-codemods",
- "@instructure/ui-color-utils",
- "@instructure/ui-date-input",
- "@instructure/ui-date-time-input",
- "@instructure/ui-decorator",
- "@instructure/ui-dialog",
- "@instructure/ui-dom-utils",
- "@instructure/ui-drawer-layout",
- "@instructure/ui-editable",
- "@instructure/ui-eslint-config",
- "@instructure/ui-expandable",
- "@instructure/ui-file-drop",
- "@instructure/ui-flex",
- "@instructure/ui-focusable",
- "@instructure/ui-form-field",
- "@instructure/ui-grid",
- "@instructure/ui-heading",
- "@instructure/ui-i18n",
- "@instructure/ui-icons-build",
- "@instructure/ui-icons",
- "@instructure/ui-img",
- "@instructure/ui-karma-config",
- "@instructure/ui-link",
- "@instructure/ui-list",
- "@instructure/ui-menu",
- "@instructure/ui-metric",
- "@instructure/ui-modal",
- "@instructure/ui-motion",
- "@instructure/ui-navigation",
- "@instructure/ui-number-input",
- "@instructure/ui-options",
- "@instructure/ui-overlays",
- "@instructure/ui-pages",
- "@instructure/ui-pagination",
- "@instructure/ui-pill",
- "@instructure/ui-popover",
- "@instructure/ui-portal",
- "@instructure/ui-position",
- "@instructure/ui-progress",
- "@instructure/ui-prop-types",
- "@instructure/ui-radio-input",
- "@instructure/ui-range-input",
- "@instructure/ui-rating",
- "@instructure/ui-react-utils",
- "@instructure/ui-responsive",
- "@instructure/ui-scripts",
- "@instructure/ui-select",
- "@instructure/ui-selectable",
- "@instructure/ui-simple-select",
- "@instructure/ui-spinner",
- "@instructure/ui-stylelint-config",
- "@instructure/ui-svg-images",
- "@instructure/ui-table",
- "@instructure/ui-tabs",
- "@instructure/ui-tag",
- "@instructure/ui-template-scripts",
- "@instructure/ui-test-locator",
- "@instructure/ui-test-queries",
- "@instructure/ui-test-sandbox",
- "@instructure/ui-test-utils",
- "@instructure/ui-testable",
- "@instructure/ui-text-area",
- "@instructure/ui-text-input",
- "@instructure/ui-text",
- "@instructure/ui-theme-tokens",
- "@instructure/ui-themes",
- "@instructure/ui-time-select",
- "@instructure/ui-toggle-details",
- "@instructure/ui-token-scripts",
- "@instructure/ui-tooltip",
- "@instructure/ui-tray",
- "@instructure/ui-tree-browser",
- "@instructure/ui-truncate-text",
- "@instructure/ui-upgrade-scripts",
- "@instructure/ui-utils",
- "@instructure/ui-view",
- "@instructure/ui-webpack-config",
- "@instructure/ui",
- "@instructure/uid"
+  "@instructure/babel-plugin-transform-imports",
+  "@instructure/browserslist-config-instui",
+  "@instructure/canvas-high-contrast-theme",
+  "@instructure/canvas-theme",
+  "@instructure/command-utils",
+  "@instructure/config-loader",
+  "@instructure/console",
+  "@instructure/cz-lerna-changelog",
+  "@instructure/debounce",
+  "@instructure/emotion",
+  "eslint-plugin-instructure-ui",
+  "@instructure/instructure-theme",
+  "@instructure/instui-cli",
+  "@instructure/instui-config",
+  "@instructure/pkg-utils",
+  "@instructure/shared-types",
+  "@instructure/template-app",
+  "@instructure/template-component",
+  "@instructure/template-package",
+  "@instructure/ui-a11y-content",
+  "@instructure/ui-a11y-utils",
+  "@instructure/ui-alerts",
+  "@instructure/ui-avatar",
+  "@instructure/ui-axe-check",
+  "@instructure/ui-babel-preset",
+  "@instructure/ui-badge",
+  "@instructure/ui-billboard",
+  "@instructure/ui-breadcrumb",
+  "@instructure/ui-buttons",
+  "@instructure/ui-byline",
+  "@instructure/ui-calendar",
+  "@instructure/ui-checkbox",
+  "@instructure/ui-code-editor",
+  "@instructure/ui-codemods",
+  "@instructure/ui-color-utils",
+  "@instructure/ui-date-input",
+  "@instructure/ui-date-time-input",
+  "@instructure/ui-decorator",
+  "@instructure/ui-dialog",
+  "@instructure/ui-dom-utils",
+  "@instructure/ui-drawer-layout",
+  "@instructure/ui-editable",
+  "@instructure/ui-eslint-config",
+  "@instructure/ui-expandable",
+  "@instructure/ui-file-drop",
+  "@instructure/ui-flex",
+  "@instructure/ui-focusable",
+  "@instructure/ui-form-field",
+  "@instructure/ui-grid",
+  "@instructure/ui-heading",
+  "@instructure/ui-i18n",
+  "@instructure/ui-icons-build",
+  "@instructure/ui-icons",
+  "@instructure/ui-img",
+  "@instructure/ui-karma-config",
+  "@instructure/ui-link",
+  "@instructure/ui-list",
+  "@instructure/ui-menu",
+  "@instructure/ui-metric",
+  "@instructure/ui-modal",
+  "@instructure/ui-motion",
+  "@instructure/ui-navigation",
+  "@instructure/ui-number-input",
+  "@instructure/ui-options",
+  "@instructure/ui-overlays",
+  "@instructure/ui-pages",
+  "@instructure/ui-pagination",
+  "@instructure/ui-pill",
+  "@instructure/ui-popover",
+  "@instructure/ui-portal",
+  "@instructure/ui-position",
+  "@instructure/ui-progress",
+  "@instructure/ui-prop-types",
+  "@instructure/ui-radio-input",
+  "@instructure/ui-range-input",
+  "@instructure/ui-rating",
+  "@instructure/ui-react-utils",
+  "@instructure/ui-responsive",
+  "@instructure/ui-scripts",
+  "@instructure/ui-select",
+  "@instructure/ui-selectable",
+  "@instructure/ui-simple-select",
+  "@instructure/ui-spinner",
+  "@instructure/ui-stylelint-config",
+  "@instructure/ui-svg-images",
+  "@instructure/ui-table",
+  "@instructure/ui-tabs",
+  "@instructure/ui-tag",
+  "@instructure/ui-template-scripts",
+  "@instructure/ui-test-locator",
+  "@instructure/ui-test-queries",
+  "@instructure/ui-test-sandbox",
+  "@instructure/ui-test-utils",
+  "@instructure/ui-testable",
+  "@instructure/ui-text-area",
+  "@instructure/ui-text-input",
+  "@instructure/ui-text",
+  "@instructure/ui-theme-tokens",
+  "@instructure/ui-themes",
+  "@instructure/ui-time-select",
+  "@instructure/ui-toggle-details",
+  "@instructure/ui-token-scripts",
+  "@instructure/ui-tooltip",
+  "@instructure/ui-tray",
+  "@instructure/ui-tree-browser",
+  "@instructure/ui-truncate-text",
+  "@instructure/ui-utils",
+  "@instructure/ui-view",
+  "@instructure/ui-webpack-config",
+  "@instructure/ui",
+  "@instructure/uid"
 ]
