--- packages/ui-codemods/__testfixtures__/deprecatePropNames/declaration.input.js
@@ -1,9 +1,0 @@
-/* eslint-disable */
-
-class Modal extends React.Component {
-  render() {
-    const foo = this.props.isOpen
-    const bar = props.isOpen
-    const { isOpen } = props
-  }
-}
