--- packages/ui-codemods/__testfixtures__/deprecatePropNames/declaration.output.js
@@ -1,9 +1,0 @@
-/* eslint-disable */
-
-class Modal extends React.Component {
-  render() {
-    const foo = this.props.open
-    const bar = props.open
-    const { open } = props
-  }
-}
