--- packages/ui-codemods/__testfixtures__/deprecatePropNames/decorator.input.js
@@ -1,13 +1,0 @@
-/* eslint-disable */
-
-@deprecated('1.0.0', {
-  foo: 'bar'
-})
-class Modal extends React.Component {}
-
-@deprecated('3.0.0', {
-  baz: 'qux'
-})
-class Popover extends React.Component {}
-
-class Position extends React.Component {}
