--- packages/ui-codemods/__testfixtures__/deprecatePropNames/defaultProps.input.js
@@ -1,7 +1,0 @@
-/* eslint-disable */
-
-class Modal extends React.Component {
-  static defaultProps = {
-    isOpen: true
-  }
-}
