--- packages/ui-codemods/__testfixtures__/deprecatePropNames/defaultProps.output.js
@@ -1,7 +1,0 @@
-/* eslint-disable */
-
-class Modal extends React.Component {
-  static defaultProps = {
-    open: true
-  }
-}
