--- packages/ui-codemods/__testfixtures__/deprecatePropNames/propMember.input.js
@@ -1,8 +1,0 @@
-/* eslint-disable */
-
-class Modal extends React.Component {
-  render() {
-    this.props.isOpen
-    props.isOpen
-  }
-}
