--- packages/ui-codemods/__testfixtures__/deprecatePropNames/propMember.output.js
@@ -1,8 +1,0 @@
-/* eslint-disable */
-
-class Modal extends React.Component {
-  render() {
-    this.props.open
-    props.open
-  }
-}
