--- packages/ui-codemods/lib/deprecatePropNames.js
@@ -1,59 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2015 - present Instructure, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-const fs = require('fs')
-const path = require('path')
-
-// const insertOrUpdateDecorator = require('./helpers/insertOrUpdateDecorator')
-const replacePropsAndDefaults = require('./helpers/replacePropsAndDefaults')
-const replaceVariableDeclarations = require('./helpers/replaceVariableDeclarations')
-const replacePropertyMembers = require('./helpers/replacePropertyMembers')
-
-module.exports = function (file, api, options) {
-  const j = api.jscodeshift
-  const c =
-    options.config.indexOf(process.cwd()) > -1
-      ? options.config
-      : path.join(process.cwd(), options.config)
-
-  const config = fs.existsSync(c) ? require(c) : null
-
-  if (!config) {
-    throw new Error(`Invalid config file "${options.config}"`)
-  }
-
-  const root = j(file.source)
-  let hasModifications = false
-
-  // TODO Modifying decorators needs some love to get it working.
-  //      In the mean time this can be done manually easily enough.
-  // hasModifications = insertOrUpdateDecorator(j, root, config) || hasModifications
-  hasModifications =
-    replacePropsAndDefaults(j, root, config) || hasModifications
-  hasModifications =
-    replaceVariableDeclarations(j, root, config) || hasModifications
-  hasModifications = replacePropertyMembers(j, root, config) || hasModifications
-
-  return hasModifications ? root.toSource({ quote: 'single' }) : null
-}
