+++ packages/ui-scripts/package.json
@@ -28,9 +28,9 @@
     "@codesandbox/common": "^1.0.8",
     "@instructure/command-utils": "8.14.0",
     "@instructure/cz-lerna-changelog": "8.14.0",
+    "@instructure/instui-config": "8.14.0",
     "@instructure/pkg-utils": "8.14.0",
     "@instructure/ui-template-scripts": "8.14.0",
-    "@instructure/ui-upgrade-scripts": "8.14.0",
     "@lerna/project": "^4",
     "commitizen": "^4.0.3",
     "dotenv": "^8.1.0",
