--- packages/ui-upgrade-scripts/.gitignore
@@ -1,1 +1,0 @@
-types/
