--- packages/ui-upgrade-scripts/lib/utils/generate-package-list.js
@@ -1,45 +1,0 @@
-/*
- * The MIT License (MIT)
- *
- * Copyright (c) 2015 - present Instructure, Inc.
- *
- * Permission is hereby granted, free of charge, to any person obtaining a copy
- * of this software and associated documentation files (the "Software"), to deal
- * in the Software without restriction, including without limitation the rights
- * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
- * copies of the Software, and to permit persons to whom the Software is
- * furnished to do so, subject to the following conditions:
- *
- * The above copyright notice and this permission notice shall be included in all
- * copies or substantial portions of the Software.
- *
- * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
- * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
- * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
- * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
- * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
- * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
- * SOFTWARE.
- */
-
-const fse = require('fs-extra')
-const path = require('path')
-const { info, error, runCommandSync } = require('@instructure/command-utils')
-
-module.exports = ({ outputDir, name = 'package-list.json' }) => {
-  try {
-    const { stdout } = runCommandSync('lerna', ['list', '--json'], [], {
-      stdio: 'pipe'
-    })
-    const packages = JSON.parse(stdout)
-    const packageList = packages.map((pkg) => pkg.name)
-
-    const outputPath = path.join(outputDir, name)
-
-    fse.outputFileSync(outputPath, JSON.stringify(packageList, null, 1))
-
-    info(`Successfully generated package list at ${outputPath}`)
-  } catch (err) {
-    error(err)
-  }
-}
