--- packages/ui-upgrade-scripts/package.json
@@ -1,34 +1,0 @@
-{
-  "name": "@instructure/ui-upgrade-scripts",
-  "version": "8.14.0",
-  "description": "A collection of UI scripts and utils to automate upgrades.",
-  "author": "Instructure, Inc. Engineering and Product Design",
-  "type": "commonjs",
-  "main": "./lib/index.js",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/instructure/instructure-ui.git"
-  },
-  "homepage": "https://instructure.github.io/instructure-ui/",
-  "bugs": "https://github.com/instructure/instructure-ui/issues",
-  "bin": {
-    "ui-upgrade-scripts": "./lib/index.js"
-  },
-  "scripts": {
-    "lint": "ui-test --lint",
-    "lint:fix": "ui-test --lint --fix"
-  },
-  "license": "MIT",
-  "dependencies": {
-    "@instructure/command-utils": "8.14.0",
-    "@instructure/pkg-utils": "8.14.0",
-    "depcheck": "^1",
-    "fs-extra": "^9",
-    "jscodeshift": "^0.7.0",
-    "yargs": "^16"
-  },
-  "publishConfig": {
-    "access": "public"
-  },
-  "sideEffects": false
-}
