--- packages/ui-upgrade-scripts/tsconfig.build.json
@@ -1,14 +1,0 @@
-{
-  "extends": "../../tsconfig.build.json",
-  "compilerOptions": {
-    "outDir": "./types",
-    "composite": true,
-    "rootDir": "./lib",
-    "module": "CommonJS"
-  },
-  "include": ["lib"],
-  "references": [
-    { "path": "../command-utils/tsconfig.build.json" },
-    { "path": "../pkg-utils/tsconfig.build.json" }
-  ]
-}
