--- packages/ui-upgrade-scripts/tsconfig.json
@@ -1,4 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {}
-}
